mvp/mvp3/requirements.txt
Flask==2.3.2
requests==2.31.0
pandas==2.1.1
numpy==1.25.0
openpyxl==3.1.2
gunicorn==21.2.0
biopython==1.81